Squashed commit of the following:
commit 242d98dc066b0b6d03bce0be05845d3077438eda
Merge: 1cfafe1 8052a42
Author: HoriYeori <xsrsx2000@gmail.com>
Date:   Wed Apr 26 02:22:58 2023 +0000

    Merge branch 'like' into 'develop'

    FEAT: add selectedCardList

    See merge request HoriYeori/dsquare-user-backend!21

commit 8052a42bd5349d3ecec73e9c06b105a6af173d98
Author: YH <49009854+100yeony@users.noreply.github.com>
Date:   Wed Apr 26 10:12:35 2023 +0900

    STYLE: code refactoring

commit 63a0ec1587fc8c7cbb1a7f53107439d08f062011
Author: YH <49009854+100yeony@users.noreply.github.com>
Date:   Tue Apr 25 17:46:53 2023 +0900

    FEAT: add selectedCardList

commit 3dec3d7a8ea9e3a328a4440cef36e3ca0a5cf158
Author: YH <49009854+100yeony@users.noreply.github.com>
Date:   Tue Apr 25 15:47:26 2023 +0900

    STYLE: like refactoring

+++ src/main/java/com/ktds/dsquare/board/qna/service/AnswerService.java
@@ -44,7 +44,7 @@ public class AnswerService {
         List<AnswerResponse> answerResponses = new ArrayList<>();
         List<Answer> answers = answerRepository.findByQuestionAndDeleteYnOrderByCreateDateAsc(qid, false);
         for(Answer answer:answers){
-            Integer likeCnt = likeService.findLikeCnt(BoardType.ANSWER, answer.getId());
+            Long likeCnt = likeService.findLikeCnt(BoardType.ANSWER, answer.getId());
             Boolean likeYn = likeService.findLikeYn(BoardType.ANSWER, answer.getId(), user);
             Long commentCnt = commentRepository.countByBoardTypeAndPostId(BoardType.ANSWER, qid.getQid());
             answerResponses.add(AnswerResponse.toDto(answer, MemberInfo.toDto(answer.getWriter()), likeCnt, likeYn, commentCnt));
